Add Array fracturing
There is a finite limit to the size of a query reddit will accept. Ruby
does not provide a mechanism for splitting an array into a series of
subsections based on length, so this commit adds one.

Array#fracture creates a set of several arrays from a source array, such
that each array in the returned set is smaller than the given maximum
length. Array#fracture preserves all other properties of the input data.