Enable the Client to request actual posts
Client#get_posts retrieves posts in batches, thanks to Array#fracture,
and yields each post to a block. It returns an array of the responses
from reddit.

This method behaves as an iterator for yielding, and the block to which
it yields will operate on a single post at a time.